Discretizations

Spatial discretization with FEM

GMRFs.shape_gradient_local — Function

shape_gradient_local(f::FEMDiscretization, shape_idx::Int, ξ)

Gradient of the shape function with index shape_idx with respect to the local coordinates ξ.

GMRFs.shape_gradient_global — Function

shape_gradient_global(f::FEMDiscretization, dof_coords, shape_idx::Int, ξ; J⁻¹ = nothing)

Gradient of the shape function with index shape_idx in a cell with node coordinates dof_coords, taken with respect to the global coordinates but computed in terms of the local coordinates ξ.

GMRFs.shape_hessian_local — Function

shape_hessian_local(f::FEMDiscretization, shape_idx::Int, ξ)

Hessian of the shape function with index shape_idx with respect to the local coordinates ξ.

GMRFs.shape_hessian_global — Function

shape_hessian_global(f::FEMDiscretization, dof_coords, shape_idx::Int, ξ; J⁻¹ = nothing, geo_hessian = nothing)

Hessian of the shape function with index shape_idx in a cell with node coordinates dof_coords, taken with respect to the global coordinates but computed in terms of the local coordinates ξ.

GMRFs.geom_jacobian — Function

geom_jacobian(f::FEMDiscretization, dof_coords, ξ)

Jacobian of the geometry mapping at the local coordinates ξ with node coordinates dof_coords. By "geometry mapping", we mean the mapping from the reference element to the physical element.

GMRFs.geom_hessian — Function

geom_hessian(f::FEMDiscretization, dof_coords, ξ)

Hessian of the geometry mapping at the local coordinates ξ with node coordinates dof_coords. By "geometry mapping", we mean the mapping from the reference element to the physical element.