Gauss-Newton Optimization

GaussianMarkovRandomFields.GaussNewtonOptimizer — Type

GaussNewtonOptimizer

Tunable Gauss-Newton optimization routine to find the maximum a posteriori (MAP) estimate under nonlinear observations and a GMRF prior. TODO: Explain parameters.

GaussianMarkovRandomFields.optimize — Function

optimize(optim::GaussNewtonOptimizer)

Iterate until the stopping criterion is fulfilled.

GaussianMarkovRandomFields.take_step — Function

take_step(optim::GaussNewtonOptimizer)

Take a Gauss-Newton step in the optimization process.

Solvers

GaussianMarkovRandomFields.GNLinearSolverBlueprint — Type

GNLinearSolverBlueprint

Abstract type for the specification of a solver for the linear systems arising in Gauss-Newton optimization.

GaussianMarkovRandomFields.GNCholeskySolverBlueprint — Type

GNCholeskySolverBlueprint(perm)

Specification of a linear solver for Gauss-Newton systems based on the Cholesky decomposition. perm is a node reordering (permutation) to minimize fill-in. If such a reordering is available from previous computations, it can be reused here to avoid unnecessary computational overhead. If perm is not passed, it will be computed during the Cholesky decomposition.

GaussianMarkovRandomFields.GNCGSolverBlueprint — Type

GNCGSolverBlueprint(; maxiter, reltol, abstol, preconditioner_fn, verbose)

Specification of a linear solver for Gauss-Newton systems based on the conjugate gradient (CG) method.

Line search strategies

GaussianMarkovRandomFields.AbstractLineSearch — Type

AbstractLineSearch

Abstract type for the specification of a line search scheme for optimization.

GaussianMarkovRandomFields.BacktrackingLineSearch — Type

BacktrackingLineSearch

Specification of a line search based on backtracking via the Armijo condition. TODO: Description of parameters

GaussianMarkovRandomFields.NoLineSearch — Type

NoLineSearch

A type that communicates that no line search will be used, i.e. the initial step proposed by the optimization algorithm is the step that will be taken.

Stopping criteria

GaussianMarkovRandomFields.AbstractStoppingCriterion — Type

AbstractStoppingCriterion

Abstract type for the specification of a criterion that tells an optimization algorithm when to stop iterating.

GaussianMarkovRandomFields.NewtonDecrementCriterion — Type

NewtonDecrementCriterion(threshold)

Stops the optimization procedure when ∇f(xₖ) ∇²f(xₖ)⁻¹ ∇f(xₖ) < threshold.

GaussianMarkovRandomFields.StepNumberCriterion — Type

StepNumberCriterion(max_steps)

Stops the optimization procedure when a maximum number of iterations / steps is reached.

GaussianMarkovRandomFields.OrCriterion — Type

OrCriterion(criteria)

Stops the optimization procedure when any of the criteria in criteria are fulfilled.